Extracting non-lsp4j functionality of navigator into own class
